assistente_financeiro/requirements.txt
# =============================================
# Dependências do Assistente Financeiro Pessoal
# =============================================

# Framework Web e API REST
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Banco de Dados ORM
sqlalchemy>=2.0.23
alembic>=1.12.1

# Validação de dados
pydantic>=2.5.0
pydantic-settings>=2.1.0

# OCR e processamento de PDF
pytesseract>=0.3.10
pdf2image>=1.16.3
PyMuPDF>=1.24.0
Pillow>=10.4.0
python-docx>=1.1.0

# Análise e manipulação de dados
pandas>=2.1.3
numpy>=1.26.2
openpyxl>=3.1.2
xlsxwriter>=3.1.9
ofxparse>=0.21

# Upload e utilitários HTTP
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.2
requests>=2.31.0

# Telegram Bot (preparado para integração futura)
python-telegram-bot>=20.6

# IA com Google Gemini (assistente Vorcaro)
google-genai>=1.0.0
openai>=1.60.0

# Geração de relatórios PDF
reportlab>=4.0.7

# Interface gráfica moderna e leve
customtkinter>=5.2.1
matplotlib>=3.8.2
